% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yarra_view.R
\name{yarra_view}
\alias{yarra_view}
\title{Launch Water Quality Dashboard}
\usage{
yarra_view()
}
\description{
Launches an interactive Shiny dashboard to explore the water quality dataset
}
\examples{
if (interactive()) {
yarra_view()
}
}
